Revert "Windows Shutdown Screen (#98)" (#100)
This reverts commit 1dcb859405119347d6800859b6b76190b61f4387.
