+++ src/components/portfolio/footer/windows-theme-button/WindowsThemeButton.tsx
@@ -2,6 +2,7 @@ import { Dispatch, SetStateAction } from 'react';
 import Image from 'next/image';
 import { Button, Tooltip } from '@mui/material';
 import ThemeConstants from '../../../../constants/theme';
+import { updateWindowThemeBgColor } from '../../../../utils/window';
 
 import imgWhite from '../../../../../public/img/windows/windows-white.png';
 import imgBlack from '../../../../../public/img/windows/windows-black.png';
@@ -21,7 +22,10 @@ export default function WindowsThemeButton({ useDarkTheme, setTheme }: Props): J
     <Tooltip title="take me back in time">
       <Button
         className=" btn-primary"
-        onClick={() => setTheme(ThemeConstants.WINDOWS)}
+        onClick={() => {
+          setTheme(ThemeConstants.WINDOWS);
+          updateWindowThemeBgColor(ThemeConstants.WINDOWS);
+        }}
         id={windowsThemeButtonId}
       >
         <Image
