+++ src/components/windows-theme/layout/Layout.tsx
@@ -3,7 +3,6 @@ import Desktop from '../desktop/Desktop';
 import StartBar from '../start-bar/StartBar';
 import DesktopItemModel from '../../../models/component-helpers/desktop-item';
 import MenuItemModel from '../../../models/component-helpers/menu-item';
-import ShutdownScreen from '../shutdown-screen/ShutdownScreen';
 
 type Props = {
   desktopItems: DesktopItemModel[];
@@ -20,7 +19,6 @@ export default function Layout({
 }: Props): JSX.Element {
   return (
     <>
-      <ShutdownScreen />
       <Desktop desktopItems={desktopItems} />
       <StartBar menuItems={menuItems} useDarkTheme={useDarkTheme} setTheme={setTheme} />
     </>
