--- src/components/windows-theme/shutdown-screen/ShutdownScreen.module.css
@@ -1,18 +1,0 @@
-.shutdown-div {
-  display: none;
-  height: 100%;
-  width: 100%;
-  z-index: 9999;
-  position: absolute;
-  background-color: #a7cce7;
-}
-
-.shutdown-img {
-  width: 100%;
-  height: auto;
-  left: 50%;
-  top: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-  position: absolute;
-}
