+++ src/components/windows-theme/start-bar/start-menu-item/StartMenuItem.tsx
@@ -2,13 +2,13 @@ import { Dispatch, SetStateAction } from 'react';
 import Image from 'next/image';
 import MenuItemModel from '../../../../models/component-helpers/menu-item';
 import ThemeConstants from '../../../../constants/theme';
-import { openCloseWindow, showHideElement } from '../../../../utils/window';
-import { shutdownScreenDivId, shutdownScreenAudioId } from '../../shutdown-screen/ShutdownScreen';
+import { openCloseMenu, openCloseWindow, updateWindowThemeBgColor } from '../../../../utils/window';
 
 import img from '../../../../../public/img/windows/ie.png';
 import styles from './StartMenuItem.module.css';
 
 type Props = {
+  startMenuDivId: string;
   showStartMenu: boolean;
   setShowStartMenu: Dispatch<SetStateAction<boolean>>;
   useDarkTheme: boolean;
@@ -21,6 +21,7 @@ export const shutdownMenuItemLabelId = 'shutdown-menu-item';
 export const shutdownLabelText = 'Shutdown';
 
 export default function StartMenuItem({
+  startMenuDivId,
   showStartMenu,
   setShowStartMenu,
   useDarkTheme,
@@ -43,23 +44,17 @@ export default function StartMenuItem({
 
   function handleWindowOpen(windowId: string, startBarButtonId: string) {
     openCloseWindow(windowId, startBarButtonId);
+    openCloseMenu(startMenuDivId, showStartMenu);
     setShowStartMenu(!showStartMenu);
   }
 
   function handleShutdown() {
     const theme = useDarkTheme ? ThemeConstants.DARK : ThemeConstants.LIGHT;
-    const shutdownScreen = document.getElementById(shutdownScreenDivId);
-    const audio = document.getElementById(shutdownScreenAudioId) as HTMLAudioElement;
 
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    if (audio) audio.play();
-    if (shutdownScreen) showHideElement(shutdownScreen);
-
+    setTheme(theme);
+    updateWindowThemeBgColor(theme);
+    openCloseMenu(startMenuDivId, showStartMenu);
     setShowStartMenu(!showStartMenu);
-
-    setTimeout(() => {
-      setTheme(theme);
-    }, 2500);
   }
 
   return (
