+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
 import Image from 'next/image';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { StyledEngineProvider } from '@mui/material/styles';
 import 'bootstrap/dist/css/bootstrap.css';
 import '../styles/globals.css';
@@ -17,7 +17,6 @@ import Window from '../components/windows-theme/window/Window';
 import PageModel from '../models/component-helpers/page';
 import DesktopItemModel from '../models/component-helpers/desktop-item';
 import ThemeConstants from '../constants/theme';
-import { updateWindowThemeBgColor } from '../utils/window';
 
 export default function App({ Component, pageProps }: AppProps) {
   const [pageValue, setPageValue] = useState<number>(0);
@@ -26,11 +25,6 @@ export default function App({ Component, pageProps }: AppProps) {
     useDarkTheme ? ThemeConstants.DARK : ThemeConstants.LIGHT
   );
   const useWindowsTheme = theme === ThemeConstants.WINDOWS;
-
-  useEffect(() => {
-    updateWindowThemeBgColor(theme);
-    document.documentElement.setAttribute('data-theme', theme);
-  }, [theme]);
 
   const handlePageChange = (event: React.SyntheticEvent, newValue: number) =>
     setPageValue(newValue);
