+++ tests/components/windows-theme/start-bar/start-button/StartButton.test.tsx
@@ -3,6 +3,7 @@ import { render } from '@testing-library/react';
 import StartButton, {
   startButtonId,
 } from '../../../../../src/components/windows-theme/start-bar/start-button/StartButton';
+import { startMenuDivId } from '../../../../../src/components/windows-theme/start-bar/StartBar';
 import { useStateMock, setBoolStateMock } from '../../../../mocks/use-state-mock';
 
 jest.mock('react', () => ({
@@ -19,7 +20,13 @@ describe('StartButton component', () => {
       .spyOn(React, 'useState')
       .mockImplementation(useStateMock as () => [unknown, Dispatch<unknown>]);
 
-    render(<StartButton showStartMenu={showStartMenu} setShowStartMenu={setShowStartMenu} />);
+    render(
+      <StartButton
+        startMenuDivId={startMenuDivId}
+        showStartMenu={showStartMenu}
+        setShowStartMenu={setShowStartMenu}
+      />
+    );
   });
 
   it('renders button', () => {
